feat: extract SensorLog DTOs to dedicated file and add log methods to ISensorService
