+++ IRasRag.Application/Services/Interfaces/ISensorService.cs
@@ -12,5 +12,10 @@ namespace IRasRag.Application.Services.Interfaces
         Task<Result> UpdateSensorAsync(Guid id, UpdateSensorDto updateDto);
         Task<Result> DeleteSensorAsync(Guid id);
         Task<Result<SensorHistoryDto>> GetSensorHistoryAsync(Guid sensorId, DateOnly date);
+        Task<Result<SensorLogDto>> CreateSensorLogAsync(Guid sensorId, CreateSensorLogDto dto);
+        Task<Result<PaginatedResult<SensorLogDto>>> GetSensorLogsAsync(
+            Guid sensorId,
+            SensorLogListRequest request
+        );
     }
 }
